Move panel.hpp to include folder
