+++ rviz_common/include/rviz_common/panel.hpp
