+++ rviz_common/src/rviz_common/panel.cpp
@@ -30,7 +30,7 @@
 
 #include <cstdio>
 
-#include "./panel.hpp"
+#include "rviz_common/panel.hpp"
 
 namespace rviz_common
 {
