+++ rviz_common/src/rviz_common/tool_properties_panel.hpp
@@ -31,7 +31,7 @@
 #ifndef RVIZ_COMMON__TOOL_PROPERTIES_PANEL_HPP_
 #define RVIZ_COMMON__TOOL_PROPERTIES_PANEL_HPP_
 
-#include "panel.hpp"
+#include "rviz_common/panel.hpp"
 
 namespace rviz_common
 {
