+++ rviz_common/src/rviz_common/visualization_frame.cpp
@@ -72,6 +72,8 @@
 
 #include "rviz_common/load_resource.hpp"
 #include "rviz_common/logging.hpp"
+#include "rviz_common/panel.hpp"
+#include "rviz_common/panel_dock_widget.hpp"
 #include "rviz_common/render_panel.hpp"
 #include "rviz_common/tool.hpp"
 #include "rviz_rendering/initialization.hpp"  // TODO(wjwwood): see if this is needed anymore
@@ -81,8 +83,6 @@
 #include "./failed_panel.hpp"
 #include "./loading_dialog.hpp"
 #include "./new_object_dialog.hpp"
-#include "./panel.hpp"
-#include "rviz_common/panel_dock_widget.hpp"
 #include "./panel_factory.hpp"
 #include "./screenshot_dialog.hpp"
 #include "./splash_screen.hpp"
